requirements.txt
Django==4.0.6
django-crispy-forms==1.14.0
django-extensions==3.2.0
